git-svn-id: http://monit.googlecode.com/svn/trunk@114 808b68a2-07de-11de-a1f0-819f45317607

+++ xml.c
@@ -126,7 +126,7 @@ static void document_head(Buffer_T *B) {
    "<localhostname>%s</localhostname>"
    "<controlfile>%s</controlfile>",
    Run.id,
-   Run.incarnation,
+   (long long)Run.incarnation,
    VERSION,
    (long)Util_getProcessUptime(Run.pidfile),
    Run.polltime,
